+++ projects/angular-sdk/src/public-api.ts
@@ -2,23 +2,23 @@
  * Public API Surface of the SDK
  */
 
-// SERVICES
-//export * from './services/hopper-events.service';
+// SERVICES (only specific elements)
+// export * from './services/hopper-events.service';
 export * from './services/hopper-proxy.service';
 
-// DIRECTIVES
-//export * from './directives/hopper-events.directive';
+// DIRECTIVES (only specific elements)
+// export * from './directives/hopper-events.directive';
 
-// COMPONENTS
+// COMPONENTS (only specific elements)
 export * from './components/cfar-offer-banner/cfar-offer-banner.component';
 export * from './components/cfar-offer-banner-large/cfar-offer-banner-large.component';
 export * from './components/cfar-offer-dialog/cfar-offer-dialog.component';
 export * from './components/cfar-exercise-dialog/cfar-exercise-dialog.component';
-//export * from './components/event-button/event-button.component';
+// export * from './components/event-button/event-button.component';
 
-// UTILIY CLASSES
-export * from './utils/dialog.utils';
+// UTILIY CLASSES (only specific elements)
 export * from './utils/api-translator.utils';
+export * from './utils/dialog.utils';
 
-// MODULES
+// MODULES (only specific elements)
 export * from './angular-sdk.module';